requirements.txt
discord.py-self
groq
SpeechRecognition
pyaudio
setuptools
pyttsx3